+++ tests/pattern.rs
@@ -1,6 +1,6 @@
 use tiny_skia::*;
 
-fn crate_triangle() -> Pixmap {
+fn crate_triangle() -> Pixmap<'static> {
     let mut canvas = Canvas::new(20, 20).unwrap();
 
     let mut paint = Paint::default();