+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - Improve clipping performance.
+- Add support for drawing to user-supplied slice.
 
 ## 0.2.0 - 2020-11-16
 ### Changed